Fix RESTful routes
