+++ config/routes.rb
@@ -1,3 +1,7 @@
 Rails.application.routes.draw do
-  resources :tasks
+  namespace :api do
+    namespace :v1 do
+      resources :tasks, only: [:index, :show, :create, :update, :destroy]
+    end
+  end
 end
